Add dropdown field support for empty rules
Dropdown fields can support optional values that are stored as empty strings or null. This adds support for the empty and notempty queries to find elements that do not or do (respectively) have a value selected.

TODO: Update the UI to hide the "one of" input when either of the empty operators is selected.